Create Fillings Jugs
Problem Description
There are N jugs on a table and each jug has a capacity C[i].

Each jug will be filled with water such that the amount of water from Jug 1 to Jug N will be in non-increasing order.

i.e. if Jug i has A[i] amount of water in it the A[i] >= A[i+1] for 1 <= i < N.

What is the maximum amount of water in total that can be poured in all the jugs?

Input format
The first line contains T, number of test cases. For each test case,

First line contains N, number of Jugs.

Second line contains N space separated integers, C[i].

Output format
For each test case print the maximum amount of water that can be poured in the jugs in a new line.

Sample Input 1
2

4

10 4 7 3

5

1 2 3 4 5

Sample Output 1
21

5

Explanation 1
In the first test case, we can pour 10 into the first jug, 4 into the second jug, 4 into the third jug (since it has non-increasing order), 3 into the fourth jug. Total of 21.

In the second test case, we can pour 1 into the first jug and thereafter, 1 into each of the jugs. Total 5.

Constraints
1 ≤ T ≤ 50

1 ≤ N ≤ 10^5

1 ≤ C[i] ≤ 10^9